useTableQuery.ts filie create qilib custom hook qilindi
