+++ admin-panel-frontend/app/[locale]/(dashboard)/(routes)/admins/page.tsx
@@ -6,10 +6,9 @@ import { Edit3Icon, EllipsisVertical, File, Trash2Icon } from "lucide-react";
 import { ColumnDef } from "@tanstack/react-table";
 import PaginationApi from "@/components/PaginationApi";
 import { Input } from "@/components/ui/input";
-import { Link, usePathname, useRouter } from "@/navigation";
+import { Link } from "@/navigation";
 import { Button } from "@/components/ui/button";
 import Admin from "@/types/admin";
-import { useSearchParams } from "next/navigation";
 import {
   Dialog,
   DialogClose,
@@ -22,28 +21,23 @@ import {
 import TableApi from "@/components/TableApi";
 import { DialogDescription } from "@radix-ui/react-dialog";
 import { useQueryClient } from "@tanstack/react-query";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { toast } from "@/components/ui/use-toast";
 import useApiQuery from "@/lib/useApiQuery";
 import AdminApi from "@/types/adminApi";
 import useApiMutation from "@/lib/useApiMutation";
 import useFileMutation from "@/lib/useFileMutation";
+import useTableQuery from "@/lib/useTableQuery";
 import { Plus } from "lucide-react";
 
 export default function Admins() {
   const t = useTranslations("admins");
   const tName = useTranslations("names");
-  const searchParams = useSearchParams();
-  const pathname = usePathname();
-  const pageFromUrl = Number(searchParams.get("page")) || 1;
-  const searchFromUrl = searchParams.get("search") || "";
-  const [page, setPage] = useState(pageFromUrl);
-  const [search, setSearch] = useState(searchFromUrl);
+  const { page, setPage, search, setSearch } = useTableQuery();
   const { data } = useApiQuery<AdminApi>(
     `admin/list?page=${[page]}&name=${search}`,
     ["admins", page, search]
   );
-  const router = useRouter();
   const queryClient = useQueryClient();
   const [adminId, setAdminId] = useState<number | null>(null);
   const { mutate } = useApiMutation<{ message: string }>(
@@ -65,19 +59,6 @@ export default function Admins() {
   const { mutate: exportAdmins } = useFileMutation(`admin/export`, [
     "exportAdmins",
   ]);
-
-  useEffect(() => {
-    const params = new URLSearchParams();
-
-    if (page > 1) {
-      params.set("page", page.toString());
-    }
-    if (search.trim() !== "") {
-      params.set("search", search);
-    }
-
-    router.replace(`${pathname}?${params.toString()}`, { scroll: false });
-  }, [page, search, pathname, router]);
 
   const adminColumns: ColumnDef<Admin>[] = [
     {
