+++ admin-panel-frontend/app/[locale]/(dashboard)/(routes)/groups/page.tsx
@@ -8,12 +8,11 @@ import PaginationApi from "@/components/PaginationApi";
 import { Input } from "@/components/ui/input";
 import Group from "@/types/group";
 import GroupApi from "@/types/groupApi";
-import { Link, usePathname, useRouter } from "@/navigation";
+import { Link } from "@/navigation";
 import { Button } from "@/components/ui/button";
 import TableApi from "@/components/TableApi";
 import { useQueryClient } from "@tanstack/react-query";
-import { useEffect, useState } from "react";
-import { useSearchParams } from "next/navigation";
+import { useState } from "react";
 import {
   Dialog,
   DialogClose,
@@ -29,29 +28,11 @@ import useApiQuery from "@/lib/useApiQuery";
 import useApiMutation from "@/lib/useApiMutation";
 import useFileMutation from "@/lib/useFileMutation";
 import { Plus } from "lucide-react";
+import useTableQuery from "@/lib/useTableQuery";
 
 export default function Groups() {
   const t = useTranslations("groups");
-  const searchParams = useSearchParams();
-  const pageFromUrl = Number(searchParams.get("page")) || 1;
-  const searchFromUrl = searchParams.get("search") || "";
-  const [page, setPage] = useState(pageFromUrl);
-  const [search, setSearch] = useState(searchFromUrl);
-  const router = useRouter();
-  const pathName = usePathname();
-
-  useEffect(() => {
-    const params = new URLSearchParams();
-
-    if (page > 1) {
-      params.set("page", page.toString());
-    }
-    if (search.trim() !== "") {
-      params.set("search", search);
-    }
-
-    router.replace(`${pathName}?${params.toString()}`, { scroll: false });
-  }, [page, search, pathName, router]);
+  const { page, setPage, search, setSearch } = useTableQuery();
 
   const { data } = useApiQuery<GroupApi>(
     `group/list?page=${page}&name=${search}`,
@@ -135,7 +116,7 @@ export default function Groups() {
       <div className="space-y-4">
         <div className="w-full flex justify-between">
           <h1 className="text-3xl w-2/4 font-bold">{t("groups")}</h1>
-          <Link href={`${pathName}/create`}>
+          <Link href={`/groups/create`}>
             <Button icon={<Plus className="h-5 w-5" />}>
               {t("creategroup")}
             </Button>
